+++ src/main/java/org/thoughtslive/jenkins/plugins/jira/service/JiraService.java
@@ -9,26 +9,13 @@ import java.util.concurrent.TimeUnit;
 
 import org.thoughtslive.jenkins.plugins.jira.Site;
 import org.thoughtslive.jenkins.plugins.jira.api.Comment;
-import org.thoughtslive.jenkins.plugins.jira.api.Comments;
 import org.thoughtslive.jenkins.plugins.jira.api.Component;
-import org.thoughtslive.jenkins.plugins.jira.api.Count;
-import org.thoughtslive.jenkins.plugins.jira.api.Issue;
 import org.thoughtslive.jenkins.plugins.jira.api.IssueLink;
 import org.thoughtslive.jenkins.plugins.jira.api.IssueLinkType;
-import org.thoughtslive.jenkins.plugins.jira.api.IssueLinkTypes;
-import org.thoughtslive.jenkins.plugins.jira.api.Issues;
-import org.thoughtslive.jenkins.plugins.jira.api.Notify;
-import org.thoughtslive.jenkins.plugins.jira.api.Project;
 import org.thoughtslive.jenkins.plugins.jira.api.ResponseData;
 import org.thoughtslive.jenkins.plugins.jira.api.SearchResult;
-import org.thoughtslive.jenkins.plugins.jira.api.Status;
-import org.thoughtslive.jenkins.plugins.jira.api.Transitions;
 import org.thoughtslive.jenkins.plugins.jira.api.User;
-import org.thoughtslive.jenkins.plugins.jira.api.Version;
-import org.thoughtslive.jenkins.plugins.jira.api.Watches;
 import org.thoughtslive.jenkins.plugins.jira.api.input.BasicIssue;
-import org.thoughtslive.jenkins.plugins.jira.api.input.BasicIssues;
-import org.thoughtslive.jenkins.plugins.jira.api.input.TransitionInput;
 import org.thoughtslive.jenkins.plugins.jira.login.SigningInterceptor;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -86,7 +73,7 @@ public class JiraService {
    * @param id component id.
    * @return component.
    */
-  public ResponseData<Component> getComponent(final String id) {
+  public ResponseData<Object> getComponent(final String id) {
     try {
       return parseResponse(jiraEndPoints.getComponent(id).execute());
     } catch (Exception e) {
@@ -100,7 +87,7 @@ public class JiraService {
    * @param component an instance of {@link Component}
    * @return created component.
    */
-  public ResponseData<Component> createComponent(final Component component) {
+  public ResponseData<Object> createComponent(final Object component) {
     try {
       return parseResponse(jiraEndPoints.createComponent(component).execute());
     } catch (Exception e) {
@@ -128,7 +115,7 @@ public class JiraService {
    * @param id component id.
    * @return count.
    */
-  public ResponseData<Count> getComponentIssueCount(final String id) {
+  public ResponseData<Object> getComponentIssueCount(final String id) {
     try {
       return parseResponse(jiraEndPoints.getComponentIssueCount(id).execute());
     } catch (Exception e) {
@@ -142,7 +129,7 @@ public class JiraService {
    * @param issueIdOrKey issue id or key.
    * @return issue.
    */
-  public ResponseData<Issue> getIssue(final String issueIdOrKey) {
+  public ResponseData<Object> getIssue(final String issueIdOrKey) {
     try {
       return parseResponse(jiraEndPoints.getIssue(issueIdOrKey).execute());
     } catch (Exception e) {
@@ -156,7 +143,7 @@ public class JiraService {
    * @param issue
    * @return issue.
    */
-  public ResponseData<BasicIssue> createIssue(final Issue issue) {
+  public ResponseData<Object> createIssue(final Object issue) {
     try {
       return parseResponse(jiraEndPoints.createIssue(issue).execute());
     } catch (Exception e) {
@@ -164,7 +151,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<BasicIssue> updateIssue(final String issueIdOrKey, final Issue issue) {
+  public ResponseData<Object> updateIssue(final String issueIdOrKey, final Object issue) {
     try {
       return parseResponse(jiraEndPoints.updateIssue(issueIdOrKey, issue).execute());
     } catch (Exception e) {
@@ -181,7 +168,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<BasicIssues> createIssues(final Issues issues) {
+  public ResponseData<Object> createIssues(final Object issues) {
     try {
       return parseResponse(jiraEndPoints.createIssues(issues).execute());
     } catch (Exception e) {
@@ -189,7 +176,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Comments> getComments(final String issueIdorKey) {
+  public ResponseData<Object> getComments(final String issueIdorKey) {
     try {
       return parseResponse(jiraEndPoints.getComments(issueIdorKey).execute());
     } catch (Exception e) {
@@ -197,7 +184,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Comment> addComment(final String issueIdorKey, final String comment) {
+  public ResponseData<Object> addComment(final String issueIdorKey, final String comment) {
     try {
       return parseResponse(jiraEndPoints
           .addComment(issueIdorKey, Comment.builder().body(comment).build()).execute());
@@ -206,18 +193,16 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Comment> updateComment(final String issueIdorKey, final String id,
-      final String comment) {
+  public ResponseData<Object> updateComment(final String issueIdorKey, final String commentId, final String comment) {
     try {
-      return parseResponse(jiraEndPoints
-          .updateComment(issueIdorKey, id, Comment.builder().id(id).body(comment).build())
-          .execute());
+      return parseResponse(
+          jiraEndPoints.updateComment(issueIdorKey, commentId, Comment.builder().body(comment).build()).execute());
     } catch (Exception e) {
       return buildErrorResponse(e);
     }
   }
 
-  public ResponseData<Comment> getComment(final String issueIdorKey, final String commentId) {
+  public ResponseData<Object> getComment(final String issueIdorKey, final String commentId) {
     try {
       return parseResponse(jiraEndPoints.getComment(issueIdorKey, commentId).execute());
     } catch (Exception e) {
@@ -225,7 +210,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Void> notifyIssue(final String issueIdorKey, final Notify notify) {
+  public ResponseData<Void> notifyIssue(final String issueIdorKey, final Object notify) {
     try {
       return parseResponse(jiraEndPoints.notifyIssue(issueIdorKey, notify).execute());
     } catch (Exception e) {
@@ -233,7 +218,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Transitions> getTransitions(final String issueIdorKey) {
+  public ResponseData<Object> getIssueTransitions(final String issueIdorKey) {
     try {
       return parseResponse(jiraEndPoints.getTransitions(issueIdorKey).execute());
     } catch (Exception e) {
@@ -241,7 +226,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Void> transitionIssue(final String idOrKey, final TransitionInput issue) {
+  public ResponseData<Void> transitionIssue(final String idOrKey, final Object issue) {
     try {
       return parseResponse(jiraEndPoints.transitionIssue(idOrKey, issue).execute());
     } catch (Exception e) {
@@ -249,7 +234,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Watches> getIssueWatches(final String issueIdorKey) {
+  public ResponseData<Object> getIssueWatches(final String issueIdorKey) {
     try {
       return parseResponse(jiraEndPoints.getIssueWatches(issueIdorKey).execute());
     } catch (Exception e) {
@@ -260,13 +245,13 @@ public class JiraService {
   public ResponseData<Void> addIssueWatcher(final String issueIdorKey, final String userName) {
     try {
       return parseResponse(jiraEndPoints
-          .addIssueWatcher(issueIdorKey, User.builder().name(userName).build()).execute());
+          .addIssueWatcher(issueIdorKey, userName).execute());
     } catch (Exception e) {
       return buildErrorResponse(e);
     }
   }
 
-  public ResponseData<SearchResult> searchIssues(final String jql, final int startAt,
+  public ResponseData<Object> searchIssues(final String jql, final int startAt,
       final int maxResults) {
     final SearchResult searchInput =
         SearchResult.builder().jql(jql).startAt(startAt).maxResults(maxResults).build();
@@ -277,7 +262,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Project[]> getProjects() {
+  public ResponseData<Object> getProjects() {
     try {
       return parseResponse(jiraEndPoints.getProjects().execute());
     } catch (Exception e) {
@@ -285,7 +270,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Project> getProject(final String projectIdOrKey) {
+  public ResponseData<Object> getProject(final String projectIdOrKey) {
     try {
       return parseResponse(jiraEndPoints.getProject(projectIdOrKey).execute());
     } catch (Exception e) {
@@ -293,7 +278,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Version[]> getProjectVersions(final String projectIdOrKey) {
+  public ResponseData<Object> getProjectVersions(final String projectIdOrKey) {
     try {
       return parseResponse(jiraEndPoints.getProjectVersions(projectIdOrKey).execute());
     } catch (Exception e) {
@@ -301,7 +286,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Component[]> getProjectComponents(final String projectIdOrKey) {
+  public ResponseData<Object> getProjectComponents(final String projectIdOrKey) {
     try {
       return parseResponse(jiraEndPoints.getProjectComponents(projectIdOrKey).execute());
     } catch (Exception e) {
@@ -309,7 +294,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Status[]> getProjectStatuses(final String projectIdOrKey) {
+  public ResponseData<Object> getProjectStatuses(final String projectIdOrKey) {
     try {
       return parseResponse(jiraEndPoints.getProjectStatuses(projectIdOrKey).execute());
     } catch (Exception e) {
@@ -317,7 +302,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Version> getVersion(final String id) {
+  public ResponseData<Object> getVersion(final String id) {
     try {
       return parseResponse(jiraEndPoints.getVersion(id).execute());
     } catch (Exception e) {
@@ -325,7 +310,7 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Version> createVersion(final Version version) {
+  public ResponseData<Object> createVersion(final Object version) {
     try {
       return parseResponse(jiraEndPoints.createVersion(version).execute());
     } catch (Exception e) {
@@ -333,15 +318,15 @@ public class JiraService {
     }
   }
 
-  public ResponseData<Void> updateVersion(final Version version) {
+  public ResponseData<Void> updateVersion(final String id, final Object version) {
     try {
-      return parseResponse(jiraEndPoints.updateVersion(version.getId(), version).execute());
+      return parseResponse(jiraEndPoints.updateVersion(id, version).execute());
     } catch (Exception e) {
       return buildErrorResponse(e);
     }
   }
 
-  public ResponseData<IssueLinkTypes> getIssueLinkTypes() {
+  public ResponseData<Object> getIssueLinkTypes() {
     try {
       return parseResponse(jiraEndPoints.getIssueLinkTypes().execute());
     } catch (Exception e) {
@@ -357,8 +342,8 @@ public class JiraService {
     }
 
     final IssueLink issueLink = IssueLink.builder().type(IssueLinkType.builder().name(name).build())
-        .comment(linkComment).inwardIssue(Issue.builder().key(inwardIssueKey).build())
-        .outwardIssue(Issue.builder().key(outwardIssueKey).build()).build();
+        .comment(linkComment).inwardIssue(BasicIssue.builder().key(inwardIssueKey).build())
+        .outwardIssue(BasicIssue.builder().key(outwardIssueKey).build()).build();
 
     try {
       return parseResponse(jiraEndPoints.createIssueLink(issueLink).execute());
@@ -366,5 +351,12 @@ public class JiraService {
       return buildErrorResponse(e);
     }
   }
-
+  
+  public ResponseData<Object> getFields() {
+    try {
+        return parseResponse(jiraEndPoints.getFields().execute());
+    } catch (Exception e) {
+        return buildErrorResponse(e);
+    }
+  }
 }