+++ src/main/java/org/thoughtslive/jenkins/plugins/jira/steps/GetIssueWatchesStep.java
@@ -8,7 +8,6 @@ import org.jenkinsci.plugins.workflow.steps.StepContext;
 import org.jenkinsci.plugins.workflow.steps.StepExecution;
 import org.kohsuke.stapler.DataBoundConstructor;
 import org.thoughtslive.jenkins.plugins.jira.api.ResponseData;
-import org.thoughtslive.jenkins.plugins.jira.api.Watches;
 import org.thoughtslive.jenkins.plugins.jira.util.JiraStepDescriptorImpl;
 import org.thoughtslive.jenkins.plugins.jira.util.JiraStepExecution;
 
@@ -48,7 +47,7 @@ public class GetIssueWatchesStep extends BasicJiraStep {
 
   }
 
-  public static class Execution extends JiraStepExecution<ResponseData<Watches>> {
+  public static class Execution extends JiraStepExecution<ResponseData<Object>> {
 
     private static final long serialVersionUID = -2511162263195215296L;
 
@@ -61,9 +60,9 @@ public class GetIssueWatchesStep extends BasicJiraStep {
     }
 
     @Override
-    protected ResponseData<Watches> run() throws Exception {
+    protected ResponseData<Object> run() throws Exception {
 
-      ResponseData<Watches> response = verifyInput();
+      ResponseData<Object> response = verifyInput();
 
       if (response == null) {
         logger.println("JIRA: Site - " + siteName + " - Querying issue watches - idOrKey: "